moved scripts
