+++ Assets/_Application/_Twin/UI/ScrollableIntField.cs
@@ -5,7 +5,7 @@ using UnityEngine.Events;
 using UnityEngine.EventSystems;
 using UnityEngine.UI;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.UI
 {
     public class ScrollableIntField : MonoBehaviour, IScrollHandler
     {
